internal: removed the limit on 64 colors in the palette
Now, if the sequence number N of a color is greater than 64,
it cannot be stored in the current 64-bit mask, then it is saved
as an `N % 64` color with an index of `N / 64`.

So, for example, the first color will be saved with the value `0b01`
with index 0 (since 1 is less than 64), and 65 will also be saved
with the value `0b01`, but already with index 1 (since 64/63 is 1)
and so on (129 color will be also `0b01` with index 3).

When comparing the color chains of the rule and the call stack,
each mask is checked, and if one does not match, then this is a
sufficient condition to say that the color chain is not in the
function call chain.

Now one mask is included in another mask, provided that the
bitwise AND is equal to the first mask and their indices are the same.

The `ColorContainer` now stores a set of masks that describe the
colors in the chain. To find out if there is a color in the set of masks,
you need to take the mask by the `Color.Index` and make a bitwise
AND with `Color.Num`, the fact of non-equality to zero will mean
that the color is in the masks, and therefore in the container.